requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
pandas>=2.0.0
gunicorn>=21.0.0
